Support `asakusafw.sdk.*` convention on Gradle.
The following convention property is also now available:
* `asakusafw.spark.version`
  * Asakusa on Spark libraries version.
